+++ Models/Subasta.cs
@@ -9,7 +9,7 @@ namespace Subastas.Models
         public int UsuarioID { get; set; }
         public string NombreProyecto { get; set; }
         public string Descripcion { get; set; }
-        public int Calificacion { get; set; }
+        public float Calificacion { get; set; }
         public bool Estatus { get; set; }
 
     }
